fix: FormatError & model parsing & android 15 support

+++ lib/features/atomic_swap/poscan/create/domain/entities/pallet_atomic_swap_base_action.dart
@@ -14,7 +14,7 @@ class PalletAtomicSwapBalanceSwapAction extends PalletAtomicSwapBaseAction {
 
 class PalletPoscanAssetsSwapTokenSwapAction extends PalletAtomicSwapBaseAction {
   final int assetId;
-  final double value;
+  final BigInt value;
 
   @override
   final bool isPoscan = true;
